+++ images/bootstrap_master_centos.sh
@@ -36,6 +36,7 @@ INIT_CLUSTER=0
 K8S_TLS=0
 PROXY_EP=0
 IMAGE_REPO=0
+INSTALL_COMPLETE=0
 
 # proxy vars for docker
 while [ $PROXY_EP -eq 0 ]; do
@@ -184,14 +185,25 @@ done
 sudo kubeadm init --config=/tmp/kubeadm-config.yaml
 
 # clean
-sudo rm -rf /tmp/etc
-sudo rm /tmp/api_lb_ep \
-    /tmp/etcd0_ip \
-    /tmp/etcd1_ip \
-    /tmp/etcd2_ip \
-    /tmp/etcd_tls.tar.gz \
-    /tmp/image_repo \
-    /tmp/k8s_tls.tar.gz \
-    /tmp/kubeadm-config.yaml \
-    /tmp/proxy_ep
+while [ $INSTALL_COMPLETE -eq 0 ]; do
+    if [ -f /tmp/install_complete ]; then
+        sudo rm -rf /tmp/etc
+        sudo rm /tmp/api_lb_ep \
+            /tmp/etcd0_ip \
+            /tmp/etcd1_ip \
+            /tmp/etcd2_ip \
+            /tmp/etcd_tls.tar.gz \
+            /tmp/image_repo \
+            /tmp/k8s_tls.tar.gz \
+            /tmp/kubeadm-config.yaml \
+            /tmp/proxy_ep
+        INSTALL_COMPLETE=1
+    else
+        echo "cluster installation not yet complete"
+        sleep 10
+    fi
+done
+
+echo "bootstrap complete"
+exit 0
